Change `find-where` signature
This update narrows down the scope of `find-where` function.

Previously, `find-where` was thought to support any arbitrary query, but it turns out to be quite hard to proper render the results, specially, when it proposes to reflect the cardinality of the entities into the final result. For example:

```clojure
(query Products
  (include :images))
```

Should return something like:

```clojure
{:id     1
 :name   "product 1"
 :images [{:id  1
           :url "http://domain/img1.jpg"}
          {:id  2
           :url "http://domain/img2.jpg"}]}
```

The current change, only accepts the `where-clause` and the associations might be passed through the `options` argument, if the `dmi` gives support to it.